Generalize checking against bad base64 data
This commit tests all base64-encoded parts of emails to push.

This generalizes the previous ad hoc method which only worked for
detecting the PNG magic header anywhere in the message.

If a email parts/payloads claim to be base64 encoded but contain
invalid characters outside the base64 alphabet, the email is never pushed
to server. The erroneous emails are NOT sent to quarantine since they
are assumed/observed to be both rare and can be detected client
side (and notably CANNOT be detected server-side as such because they
break the protocol contract).